# mint-ensemble-manager/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mint Ensemble Manager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://portals.tapis.io/v3/oauth2/authorize",
            "scopes": {}
          }
        }
      }
    },
    "schemas": {
      "TimePeriod": {
        "type": "object",
        "description": "A time period",
        "properties": {
          "start_date": {
            "description": "From date YYYY-MM-DD format",
            "type": "string",
            "format": "date-time"
          },
          "end_date": {
            "description": "to date YYYY-MM-DD format",
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MintEvent": {
        "type": "object",
        "properties": {
          "event": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "ADD_TASK",
              "DELETE_TASK"
            ]
          },
          "userid": {
            "type": "string",
            "example": "*"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
      "Task": {
        "type": "object",
        "description": "A Task definition",
        "properties": {
          "id": {
            "description": "The id of the task (leave empty if creating a new task)",
            "type": "string"
          },
          "name": {
            "description": "The name of the task",
            "type": "string"
          },
          "indicatorid": {
            "description": "The SVO label for the indicator/response (link to list upcoming). Example: grain~dry__mass-per-area_yield",
            "type": "string",
            "example": "crop__potential_transpiration_volume_flux"
          },
          "interventionid": {
            "description": "The SVO label for the intervention (if any). Example: crop__planting_start_time",
            "type": "string",
            "example": "nitrogen__average_of_net_mass_mineralization_rate"
          },
          "regionid": {
            "description": "The specific region id for the task (browse https://dev.mint.isi.edu/ethiopia/regions to get the ids)",
            "type": "string"
          },
          "dates": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "events": {
            "type": "array",
            "description": "List of events associated with the task",
            "items": {
              "$ref": "#/components/schemas/MintEvent"
            }
          },
          "permissions": {
            "type": "array",
            "description": "List of permissions for the task",
            "items": {
              "$ref": "#/components/schemas/MintPermission"
            }
          }
        }
      },
...
      "CreateTaskAndSubtaskRequest": {
        "type": "object",
        "description": "A request to create both a task and subtask in a single operation",
        "required": [
          "task",
          "subtask"
        ],
        "properties": {
          "task": {
            "type": "object",
            "description": "The task configuration",
            "required": [
              "name",
              "response_variables",
              "driving_variables"
            ],
            "properties": {
              "name": {
                "description": "The name of the task",
                "type": "string",
                "example": "Crop Yield Analysis Task"
              },
              "response_variables": {
                "type": "array",
                "description": "List of response variables",
                "items": {
                  "type": "string"
                },
                "example": [
                  "crop__potential_transpiration_volume_flux"
                ]
              },
              "driving_variables": {
                "type": "array",
                "description": "List of driving variables",
                "items": {
                  "type": "string"
                },
                "example": [
                  "nitrogen__average_of_net_mass_mineralization_rate"
                ]
              },
              "regionid": {
                "description": "The specific region id for the task",
                "type": "string",
                "example": "ethiopia"
              },
              "dates": {
                "type": "object",
                "description": "The date range for the task",
                "properties": {
                  "start_date": {
                    "description": "Start date in ISO format",
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-01T00:00:00Z"
                  },
                  "end_date": {
                    "description": "End date in ISO format",
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-12-31T23:59:59Z"
                  }
                }
              }
            }
          },
          "subtask": {
            "type": "object",
            "description": "The subtask configuration",
            "required": [
              "name",
              "driving_variables",
              "response_variables"
            ],
            "properties": {
              "name": {
                "description": "The name of the subtask",
                "type": "string",
                "example": "Crop Yield Analysis Subtask"
              },
              "driving_variables": {
                "type": "array",
                "description": "List of driving variables",
                "items": {
                  "type": "string"
                },
                "example": [
                  "nitrogen__average_of_net_mass_mineralization_rate"
                ]
              },
              "response_variables": {
                "type": "array",
                "description": "List of response variables",
                "items": {
                  "type": "string"
                },
                "example": [
                  "crop__potential_transpiration_volume_flux"
                ]
              },
              "dates": {
                "type": "object",
                "description": "The date range for the subtask",
                "properties": {
                  "start_date": {
                    "description": "Start date in ISO format",
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-01T00:00:00Z"
                  },
                  "end_date": {
                    "description": "End date in ISO format",
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-12-31T23:59:59Z"
                  }
                }
              }
            }
          }
        },
        "example": {
          "task": {
            "name": "Crop Yield Analysis Task",
            "response_variables": [
              "crop__potential_transpiration_volume_flux"
            ],
            "driving_variables": [
              "nitrogen__average_of_net_mass_mineralization_rate"
            ],
            "regionid": "ethiopia",
            "dates": {
              "start_date": "2024-01-01T00:00:00Z",
              "end_date": "2024-12-31T23:59:59Z"
            }
          },
          "subtask": {
            "name": "Crop Yield Analysis Subtask",
            "driving_variables": [
              "nitrogen__average_of_net_mass_mineralization_rate"
            ],
            "response_variables": [
              "crop__potential_transpiration_volume_flux"
            ],
            "dates": {
              "start_date": "2024-01-01T00:00:00Z",
              "end_date": "2024-12-31T23:59:59Z"
            }
          }
        }
      },
...
      "CreateProblemStatementRequest": {
        "type": "object",
        "description": "A Problem Statement Info definition",
        "required": [
          "name",
          "regionid",
          "dates"
        ],
        "properties": {
          "name": {
            "description": "The name of the problem statement",
            "type": "string"
          },
          "regionid": {
            "description": "The top level region id of the problem statement (example south_sudan, ethiopia)",
            "type": "string",
            "example": "texas"
          },
          "dates": {
            "type": "object",
            "description": "The date range for the problem statement",
            "required": [
              "start_date",
              "end_date"
            ],
            "properties": {
              "start_date": {
                "description": "Start date in ISO format",
                "type": "string",
                "format": "date-time"
              },
              "end_date": {
                "description": "End date in ISO format",
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "events": {
            "type": "array",
            "description": "List of events associated with the problem statement",
            "items": {
              "$ref": "#/components/schemas/MintEvent"
            }
          },
          "permissions": {
            "type": "array",
            "description": "List of permissions for the problem statement",
            "items": {
              "type": "object",
              "properties": {
                "userid": {
                  "type": "string",
                  "example": "*"
                },
                "read": {
                  "type": "boolean"
                },
                "write": {
                  "type": "boolean"
                },
                "execute": {
                  "type": "boolean"
                },
                "owner": {
                  "type": "boolean"
                }
              }
            }
          },
          "preview": {
            "type": "array",
            "description": "Preview data for the problem statement",
            "items": {
              "type": "string"
            }
          }
        },
        "example": {
          "name": "Texas Water Management Analysis 2024",
          "regionid": "texas",
          "dates": {
            "start_date": "2024-01-01T00:00:00Z",
            "end_date": "2024-12-31T23:59:59Z"
          },
          "events": [
            {
              "event": "CREATE",
              "userid": "mint@isi.edu",
              "timestamp": "2024-01-15T10:30:00Z",
              "notes": "Initial problem statement creation for Texas water management analysis"
            }
          ],
          "permissions": [
            {
              "userid": "*",
              "read": true,
              "write": false,
              "execute": false,
              "owner": false
            }
          ],
          "preview": [
            "Water availability assessment for major Texas regions",
            "Climate impact analysis on water resources",
            "Agricultural water demand projections"
          ]
        }
      },
      "ProblemStatement": {
        "type": "object",
        "description": "A Problem Statement definition",
        "properties": {
          "id": {
            "description": "The id of the problem statement (leave empty if creating a new problem statement)",
            "type": "string"
          },
          "name": {
            "description": "The name of the problem statement",
            "type": "string"
          },
          "regionid": {
            "description": "The top level region id of the problem statement (example south_sudan, ethiopia)",
            "type": "string"
          },
          "dates": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "events": {
            "type": "array",
            "description": "List of events associated with the problem statement",
            "items": {
              "$ref": "#/components/schemas/MintEvent"
            }
          },
          "permissions": {
            "type": "array",
            "description": "List of permissions for the problem statement",
            "items": {
              "$ref": "#/components/schemas/MintPermission"
            }
          },
          "tasks": {
            "type": "array",
            "description": "List of tasks associated with the problem statement",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/problemStatements": {
      "get": {
        "summary": "Get all problem statements",
        "description": "Returns a list of all problem statements",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Problem Statements"
        ],
        "responses": {
          "200": {
            "description": "A list of problem statements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProblemStatement"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new problem statement",
        "description": "Creates a new problem statement with the provided information",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Problem Statements"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProblemStatementRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Problem statement created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The ID of the created problem statement"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/problemStatements/{id}": {
      "get": {
        "summary": "Get a specific problem statement",
        "description": "Returns a specific problem statement by ID",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Problem Statements"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The problem statement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemStatement"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/problemStatements/{id}/taskAndSubtask": {
      "post": {
        "summary": "Create a task and subtask together",
        "description": "Creates both a task and its associated subtask in a single operation",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Problem Statements"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskAndSubtaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task and subtask created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "taskId": {
                      "type": "string",
                      "description": "The ID of the created task"
                    },
                    "subtaskId": {
                      "type": "string",
                      "description": "The ID of the created subtask"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/problemStatements/{problemStatementId}/tasks": {
      "get": {
        "summary": "Get all tasks for a problem statement",
        "description": "Returns a list of all tasks associated with a specific problem statement",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "problemStatementId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "description": "Creates a new task for a specific problem statement",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "problemStatementId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}